+++ workspaces/ai-integrations/plugins/catalog-backend-module-model-catalog/src/providers/index.ts
@@ -14,3 +14,5 @@
  * limitations under the License.
  */
 export { ModelCatalogResourceEntityProvider } from './ModelCatalogResourceEntityProvider';
+export * from './types';
+export * from './config';