+++ workspaces/orchestrator/plugins/orchestrator-common/src/index.ts
@@ -21,6 +21,6 @@ export * from './constants';
 export * from './models';
 export * from './workflow';
 export * from './QueryParams';
-export * from './utils/StringUtils';
+export * from './utils';
 export * from './permissions';
 export * from './auth';