+++ workspaces/orchestrator/plugins/orchestrator/src/api/OrchestratorClient.ts
@@ -36,6 +36,7 @@ import {
   PaginationInfoDTO,
   ProcessInstanceDTO,
   ProcessInstanceListResultDTO,
+  RetriggerInstanceRequestDTO,
   WorkflowOverviewDTO,
   WorkflowOverviewListResultDTO,
 } from '@red-hat-developer-hub/backstage-plugin-orchestrator-common';
@@ -103,6 +104,7 @@ export class OrchestratorClient implements OrchestratorApi {
     workflowId: string;
     parameters: JsonObject;
     authTokens: AuthToken[];
+    targetEntity?: string;
   }): Promise<AxiosResponse<ExecuteWorkflowResponseDTO>> {
     const defaultApi = await this.getDefaultAPI();
     const reqConfigOption: AxiosRequestConfig =
@@ -111,6 +113,7 @@ export class OrchestratorClient implements OrchestratorApi {
     const requestBody: ExecuteWorkflowRequestDTO = {
       inputData: args.parameters,
       authTokens: args.authTokens,
+      targetEntity: args.targetEntity,
     };
     try {
       return await defaultApi.executeWorkflow(
@@ -139,14 +142,22 @@ export class OrchestratorClient implements OrchestratorApi {
   async retriggerInstance(
     workflowId: string,
     instanceId: string,
+    authTokens?: AuthToken[],
   ): Promise<AxiosResponse<object>> {
     const defaultApi = await this.getDefaultAPI();
     const reqConfigOption: AxiosRequestConfig =
       await this.getDefaultReqConfig();
+
+    const requestBody: RetriggerInstanceRequestDTO = {};
+    if (authTokens) {
+      requestBody.authTokens = authTokens;
+    }
+
     try {
       return await defaultApi.retriggerInstance(
         workflowId,
         instanceId,
+        requestBody,
         reqConfigOption,
       );
     } catch (err) {
@@ -179,6 +190,22 @@ export class OrchestratorClient implements OrchestratorApi {
     try {
       return await defaultApi.getWorkflowsOverview(
         { paginationInfo, filters },
+        reqConfigOption,
+      );
+    } catch (err) {
+      throw getError(err);
+    }
+  }
+  async getWorkflowsOverviewForEntity(
+    targetEntity: string,
+    annotationWorkflowIds: string[],
+  ): Promise<AxiosResponse<WorkflowOverviewListResultDTO>> {
+    const defaultApi = await this.getDefaultAPI();
+    const reqConfigOption: AxiosRequestConfig =
+      await this.getDefaultReqConfig();
+    try {
+      return await defaultApi.getWorkflowsOverviewForEntity(
+        { targetEntity, annotationWorkflowIds },
         reqConfigOption,
       );
     } catch (err) {