+++ workspaces/orchestrator/plugins/orchestrator/src/api/api.ts
@@ -37,6 +37,7 @@ export interface OrchestratorApi {
   retriggerInstance(
     workflowId: string,
     instanceId: string,
+    authTokens?: AuthToken[],
   ): Promise<AxiosResponse<object>>;
 
   executeWorkflow(args: {
@@ -44,6 +45,7 @@ export interface OrchestratorApi {
     parameters: JsonObject;
     authTokens: AuthToken[];
     businessKey?: string;
+    targetEntity?: string;
   }): Promise<AxiosResponse<ExecuteWorkflowResponseDTO>>;
 
   getWorkflowSource(workflowId: string): Promise<AxiosResponse<string>>;
@@ -61,9 +63,15 @@ export interface OrchestratorApi {
 
   pingWorkflowService(workflowId: string): Promise<AxiosResponse<boolean>>;
 
-  listWorkflowOverviews(): Promise<
-    AxiosResponse<WorkflowOverviewListResultDTO>
-  >;
+  listWorkflowOverviews(
+    paginationInfo?: PaginationInfoDTO,
+    filters?: Filter,
+  ): Promise<AxiosResponse<WorkflowOverviewListResultDTO>>;
+
+  getWorkflowsOverviewForEntity(
+    targetEntity: string,
+    annotationWorkflowIds: string[],
+  ): Promise<AxiosResponse<WorkflowOverviewListResultDTO>>;
 
   listInstances(
     paginationInfo?: PaginationInfoDTO,