+++ workspaces/orchestrator/plugins/orchestrator-form-widgets/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets
 
+## 1.0.3
+
+### Patch Changes
+
+- fac94ef: fix(orchestrator): on retrigger workflow, tokens requested by the AuthRequester are forwarded
+- Updated dependencies [fac94ef]
+- Updated dependencies [16439ad]
+- Updated dependencies [4fd43f1]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@3.0.3
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-form-api@2.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- 8d89f18: Remove default React imports.
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-form-api@2.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- @red-hat-developer-hub/backstage-plugin-orchestrator-form-api@2.0.1
+
 ## 1.0.0
 
 ### Major Changes